#19567 migrated the HTTP client from Netty 3 to 4 and invalidated the approach
here, so this is a port rather than a conflict resolution.

The detection read the body in handleResponse via response.getContent(). Under
Netty 4 the initial HttpResponse carries no body at all. handleResponse now
records only the status and Content-Type, and the non-JSON prefix check runs in
handleChunk against the first chunk carrying a non-whitespace byte, using the
recorded Content-Type. That call site already existed for chunked replies; it
is now the only one, which is simpler than what it replaced.

ChannelBuffer becomes ByteBuf in bodyPrefixByte, failIfNonJsonBody and
throwForNonJsonBody. All three read readerIndex/readableBytes/getByte, which
are identical on Netty 4, so only the type changed.

Dropped my NettyHttpClient change in favour of upstream's. #19567 solved the
same problem independently, with `if (!retVal.isDone()) retVal.setException(ex)`
plus a rethrow instead of routing through handleExceptionAndCloseChannel. Their
version is the codebase's now, and carrying two spellings of one fix would be
worse than taking theirs. NettyHttpClientTest is kept and ported, because
upstream added no test for that path and these two assert the behaviour rather
than the mechanism: they pass unmodified against their implementation.

DirectDruidClientTest's fake client no longer stuffs bodies[0] into the
response. Every element is delivered as an HttpContent and chunkNum starts at
0, which is what Netty 4 actually does.

Verified with JAVA_HOME=openjdk@26, mvn -pl server -am ... test:

  DirectDruidClientTest   22 passed, 0 failures  (upstream/master baseline: 12)
  JsonParserIteratorTest  16 passed, 0 failures
  NettyHttpClientTest      2 passed, 0 failures

The baseline was measured in a separate worktree at 51e31a68c2 with the
identical maven invocation. An earlier run using `surefire:test` instead of
`test` reported 14 failures on both branches; that was a stale druid-processing
jar from ~/.m2 being resolved instead of the reactor build, not a real failure.

+++ extensions-core/lookups-cached-global/src/test/java/org/apache/druid/data/input/MapPopulatorTest.java
@@ -39,7 +39,7 @@ public class MapPopulatorTest
   public void test_getByteLengthOfObject_string_stringLength()
   {
     String o = "string";
-    Assertions.assertEquals((o.length() * Character.BYTES) + 40, MapPopulator.getByteLengthOfObject(o));
+    Assertions.assertEquals(((long) o.length() * Character.BYTES) + 40, MapPopulator.getByteLengthOfObject(o));
   }
 
   @Test
